+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [0.0.2]
+
+- Added "Project Notes: Open Existing Note (Edit Mode)" Command
+- Added "Project Notes: Open Existing Note (Preview Mode)" Command
+
 ## [0.0.1]
 
 - Initial release